Move chat chunk stream to chat_models/base
